+++ workalendar/usa/core.py
@@ -59,7 +59,8 @@ class UnitedStates(WesternCalendar, ChristianMixin):
     national_memorial_day_label = "Memorial Day"
 
     # Some regional variants
-    include_mardi_gras = False
+    include_fat_tuesday = False
+    fat_tuesday_label = "Mardi Gras"
 
     # Shift day mechanism
     # These days won't be shifted to next MON or previous FRI
@@ -272,13 +273,6 @@ class UnitedStates(WesternCalendar, ChristianMixin):
             self.national_memorial_day_label
         )
 
-    def get_mardi_gras(self, year):
-        """
-        Mardi Gras is the Tuesday happening 47 days before Easter
-        """
-        sunday = self.get_easter_sunday(year)
-        return (sunday - timedelta(days=47), "Mardi Gras")
-
     def get_variable_days(self, year):
         # usual variable days
         days = super().get_variable_days(year)
@@ -294,9 +288,6 @@ class UnitedStates(WesternCalendar, ChristianMixin):
             (UnitedStates.get_nth_weekday_in_month(year, 11, THU, 4),
                 "Thanksgiving Day"),
         ])
-
-        if self.include_mardi_gras:
-            days.append(self.get_mardi_gras(year))
 
         if self.include_federal_presidents_day:
             days.append(self.get_presidents_day(year))