+++ Changelog.md
@@ -3,6 +3,7 @@
 ## master (unreleased)
 
 - Small fixes (docstrings, use of extends, etc) on Cayman Islands calendar (#507).
+- Moving Carnaval / Mardi Gras / Fat Tuesday calculation into the `workalendar.core` module, because it's used in at least 3 countries and some States / Counties in the USA.
 
 ## v10.0.0 (2020-06-05)